Remove unnneeded/deprecated stuff, turn on more warnings

+++ Sources/TableView/NSIndexPath+FKAdditions.m
@@ -7,7 +7,7 @@ FKLoadCategory(NSIndexPathFKAdditions);
 + (NSArray *)indexPathsFromRow:(NSInteger)startRow toRow:(NSInteger)endRow inSection:(NSInteger)section {
     FKAssert(endRow >= startRow, @"endRow must be greater or equal than startRow");
     
-    NSMutableArray *indexPaths = [NSMutableArray arrayWithCapacity:endRow-startRow+1];
+    NSMutableArray *indexPaths = [NSMutableArray arrayWithCapacity:(NSUInteger)(endRow-startRow+1)];
     
     for (NSInteger row = startRow; row <= endRow; row++) {
         [indexPaths addObject:[NSIndexPath indexPathForRow:row inSection:section]];
